vision/requirements.txt
opencv-python>=4.10.0
numpy>=1.26.0
pusher>=3.3.2
python-dotenv>=1.0.1
